Major Refactoring
move to glam since that's what bevy uses, store data as f32 since it's the default, start setting up PandaObject enum for storing all nodes, make documentation more robust

+++ crates/panda3d/src/nodes/color_attrib.rs
@@ -13,7 +13,7 @@ pub(crate) enum ColorType {
 #[allow(dead_code)]
 pub(crate) struct ColorAttrib {
     color_type: ColorType,
-    color: [f64; 4],
+    color: [f32; 4],
 }
 
 impl ColorAttrib {
@@ -40,10 +40,10 @@ impl ColorAttrib {
             }
             ColorType::Flat => {
                 //TODO: SIMD? once it's stabilized
-                self.color[0] = f64::floor(self.color[0] * 1024.0 + 0.5) / 1024.0;
-                self.color[1] = f64::floor(self.color[1] * 1024.0 + 0.5) / 1024.0;
-                self.color[2] = f64::floor(self.color[2] * 1024.0 + 0.5) / 1024.0;
-                self.color[3] = f64::floor(self.color[3] * 1024.0 + 0.5) / 1024.0;
+                self.color[0] = f32::floor(self.color[0] * 1024.0 + 0.5) / 1024.0;
+                self.color[1] = f32::floor(self.color[1] * 1024.0 + 0.5) / 1024.0;
+                self.color[2] = f32::floor(self.color[2] * 1024.0 + 0.5) / 1024.0;
+                self.color[3] = f32::floor(self.color[3] * 1024.0 + 0.5) / 1024.0;
             }
             ColorType::Off => {
                 self.color = [1.0, 1.0, 1.0, 1.0];
